exploring a simpler and more international-friendly geo model
* create a table for locations with a GIS latlng field and a user-inputted address field
* use google geocoder formfield to transform user-inputted address into a latlng
** but preserve the user-inputted address in the resulting cleaned data, also
* add foreignkeys to this new location table in user profiles and groups
** use this instead of zipcode

The idea is that this can replace the entire geo.Location table (as well as model-specific address components)
and the latlng data will be sufficient to display map interfaces for searching things near a place (using google maps
and bounding box queries with the map viewpane as the bbox)

Instead of attempting to store normalized location data in our database, we'd just store the latlng and the freeform
address just as the user inputted it.  More structured location data can be pulled from Google as needed; this seems
like it's not necessary for the localpower django project, though -- we can just pull that data from Google every so
often and push it over to ActionKit, so that admins can use that data in the CRM.

Note this requires a change to the local_settings (use contrib.gis DB_BACKEND) and GEOS must be installed.